Post Round 1 of 2023-2024
Adding my solutions for the Round 1 of 2023-2024 and tweaking the Maraton task a bit to support long longs as the answer requires that

+++ 2022-2023/Round 1/maraton.cpp
@@ -1,12 +1,12 @@
 #include <iostream>
 
 int main() {
-	int x = 0;
+	long long x = 0;
 	int n;
 	std::cin >> n;
 
 	for (int m = 0; m < n; m++) {
-		int addition;
+		long long addition;
 		std::cin >> addition;
 		x += addition;
 	}
